+++ dsptoolbox/room_acoustics/__init__.py
@@ -10,11 +10,12 @@ Here are contained some functions that are related to room acoustics.
 
 """
 from .room_acoustics import (reverb_time, find_modes, convolve_rir_on_signal,
-                             find_ir_start)
+                             find_ir_start, generate_synthetic_rir)
 
 __all__ = [
     'reverb_time',
     'find_modes',
     'convolve_rir_on_signal',
     'find_ir_start',
+    'generate_synthetic_rir',
 ]