+++ dsptoolbox/transfer_functions/transfer_functions.py
@@ -2,16 +2,16 @@
 Methods used for acquiring and windowing transfer functions
 """
 import numpy as np
-from dsptoolbox import Signal
+from dsptoolbox.classes.signal_class import Signal
 from dsptoolbox._general_helpers import (_find_frequencies_above_threshold)
 from ._transfer_functions import (_spectral_deconvolve,
                                   _window_this_ir)
-from dsptoolbox._standard import _welch
-from dsptoolbox.standard_functions import pad_trim
+from dsptoolbox._standard import (
+    _welch, _minimum_phase, _group_delay_direct, _pad_trim)
 
 
 def spectral_deconvolve(num: Signal, denum: Signal,
-                        mode='regularized', start_stop_hz=None,
+                        mode: str = 'regularized', start_stop_hz=None,
                         threshold_db=-30, padding: bool = False,
                         keep_original_length: bool = False) -> Signal:
     """Deconvolution by spectral division of two signals. If the denominator
@@ -32,7 +32,8 @@ def spectral_deconvolve(num: Signal, denum: Signal,
         `'automatic'` uses a threshold dBFS to create a spectral
         window for the numerator or regularized inversion. Array of 2 or
         4 frequency points can be also manually given. `None` uses no
-        spectral window.
+        spectral window. If mode is standard, start_stop_hz has to be set
+        to `None`. Default: `None`.
     threshold : int, optional
         Threshold in dBFS for the automatic creation of the window.
         Default: -30.
@@ -60,14 +61,20 @@ def spectral_deconvolve(num: Signal, denum: Signal,
         multichannel = True
     assert num.sampling_rate_hz == denum.sampling_rate_hz, \
         'Sampling rates do not match'
+    mode = mode.lower()
     assert mode in ('regularized', 'window', 'standard'),\
         f'{mode} is not supported. Use regularized, window or None'
+    if mode == 'standard':
+        assert start_stop_hz is None, \
+            'No start_stop_hz vector can be passed when using standard mode'
 
     original_length = num.time_data.shape[0]
 
     if padding:
-        num = pad_trim(num, original_length*2)
-        denum = pad_trim(denum, original_length*2)
+        num.time_data = _pad_trim(num.time_data, original_length*2)
+        denum.time_data = _pad_trim(denum.time_data, original_length*2)
+        # num = pad_trim(num, original_length*2)
+        # denum = pad_trim(denum, original_length*2)
 
     denum.set_spectrum_parameters(method='standard')
     _, denum_fft = denum.get_spectrum()
@@ -110,7 +117,8 @@ def spectral_deconvolve(num: Signal, denum: Signal,
                      signal_type='ir')
     if padding:
         if keep_original_length:
-            new_sig = pad_trim(new_sig, original_length)
+            new_sig.time_data = _pad_trim(new_sig.time_data, original_length)
+            # new_sig = pad_trim(new_sig, original_length)
     return new_sig
 
 
@@ -267,3 +275,168 @@ def compute_transfer_function(output: Signal, input: Signal, mode='h2',
                 signal_type=mode.lower())
     tf.set_coherence(coherence)
     return tf
+
+
+def spectral_average(signal: Signal):
+    """Averages all channels of a given IR using their magnitude and
+    phase spectra.
+
+    Parameters
+    ----------
+    signal : `Signal`
+        Signal with channels to be averaged over.
+
+    Returns
+    -------
+    avg_sig : `Signal`
+        Averaged signal.
+
+    """
+    assert signal.signal_type in ('rir', 'ir'), \
+        'Averaging is valid for signal types rir or ir and not ' +\
+        f'{signal.signal_type}'
+    assert signal.number_of_channels > 1, \
+        'Signal has only one channel so no meaningful averaging can be done'
+
+    # Obtain channel magnitude and phase spectra
+    _, sp = signal.get_spectrum()
+    mag = np.abs(sp)
+    pha = np.unwrap(np.angle(sp), axis=0)
+
+    # Build averages
+    new_mag = np.mean(mag, axis=1)
+    new_pha = np.mean(pha, axis=1)
+    # New signal
+    new_sp = new_mag * np.exp(1j*new_pha)
+
+    # New time data and signal object
+    new_time_data = np.fft.irfft(new_sp[..., None], axis=0)
+    avg_sig = signal.copy()
+    avg_sig.time_data = new_time_data
+    return avg_sig
+
+
+def min_phase_from_mag(spectrum: np.ndarray, sampling_rate_hz: int,
+                       signal_type: str = 'ir'):
+    """Returns a minimal phase signal from a magnitude spectrum using
+    the hilbert transform.
+
+    Parameters
+    ----------
+    spectrum : `np.ndarray`
+        Spectrum with only positive frequencies and 0.
+    sampling_rate_hz : int
+        Signal's sampling rate in Hz.
+    signal_type : str, optional
+        Type of signal to be returned. Default: `'ir'`.
+
+    Returns
+    -------
+    sig_min_phase : `Signal`
+        Signal with same magnitude spectrum but minimal phase.
+
+    References
+    ----------
+    - https://en.wikipedia.org/wiki/Minimum_phase
+
+    """
+    if spectrum.ndim < 2:
+        spectrum = spectrum[..., None]
+    assert spectrum.ndim < 3, \
+        'Spectrum should have shape (bins, channels)'
+    if spectrum.shape[0] < spectrum.shape[1]:
+        spectrum = spectrum.T
+    spectrum = np.abs(spectrum)
+    min_spectrum = np.empty(spectrum.shape, dtype='cfloat')
+    for n in range(spectrum.shape[1]):
+        phase = _minimum_phase(spectrum[:, n], False)
+        min_spectrum[:, n] = spectrum[:, n]*np.exp(1j*phase)
+    time_data = np.fft.irfft(min_spectrum, axis=0)
+    sig_min_phase = Signal(
+        None, time_data=time_data,
+        sampling_rate_hz=sampling_rate_hz, signal_type=signal_type)
+    return sig_min_phase
+
+
+def lin_phase_from_mag(spectrum: np.ndarray, sampling_rate_hz: int,
+                       group_delay_ms='minimal',
+                       check_causality: bool = True,
+                       signal_type: str = 'ir'):
+    """Returns a linear phase signal from a magnitude spectrum. It is possible
+    to return the smallest causal group delay by checking the minimal phase
+    version of the signal and choosing a constant group delay that is never
+    lower than minimum group delay (for each channel). A value for the group
+    delay can be also passed directly and applied to all channels. If check
+    causility is activated, it is assessed that the given group delay is not
+    less than each minimal group delay. If deactivated, the generated phase
+    could lead to a non-causal system!
+
+    Parameters
+    ----------
+    spectrum : `np.ndarray`
+        Spectrum with only positive frequencies and 0.
+    sampling_rate_hz : int
+        Signal's sampling rate in Hz.
+    group_delay_ms : str or float, optional
+        Constant group delay that the phase should have for all channels
+        (in ms). Pass `'minimal'` to create a signal with the minimum linear
+        phase possible (that is different for each channel).
+        Default: `'minimal'`.
+    check_causality : bool, optional
+        When `True`, it is assessed for each channel that the given group
+        delay is not lower than the minimal group delay. Default: `True`.
+    signal_type : str, optional
+        Type of signal to be returned. Default: `'ir'`.
+
+    Returns
+    -------
+    sig_lin_phase : `Signal`
+        Signal with same magnitude spectrum but linear phase.
+
+    """
+    # Check spectrum
+    if spectrum.ndim < 2:
+        spectrum = spectrum[..., None]
+    assert spectrum.ndim < 3, \
+        'Spectrum should have shape (bins, channels)'
+    if spectrum.shape[0] < spectrum.shape[1]:
+        spectrum = spectrum.T
+    spectrum = np.abs(spectrum)
+
+    # Check group delay ms parameter
+    minimum_group_delay = False
+    if type(group_delay_ms) == str:
+        group_delay_ms = group_delay_ms.lower()
+        assert group_delay_ms == 'minimal', \
+            'Group delay should be set to minimal'
+        minimum_group_delay = True
+    elif type(group_delay_ms) in (float, int):
+        group_delay_ms /= 1000
+    else:
+        raise TypeError('group_delay_ms must be either str, float or int')
+
+    # Frequency vector
+    f_vec = np.fft.rfftfreq(spectrum.shape[0]*2-1, 1/sampling_rate_hz)
+    delta_f = f_vec[1]-f_vec[0]
+
+    # New spectrum
+    lin_spectrum = np.empty(spectrum.shape, dtype='cfloat')
+    for n in range(spectrum.shape[1]):
+        if check_causality or minimum_group_delay:
+            min_phase = _minimum_phase(spectrum[:, n], False)
+            min_gd = _group_delay_direct(min_phase, delta_f)
+            gd = np.max(min_gd) + 1e-3  # add 1 ms as safety factor
+            if check_causality and type(group_delay_ms) != str:
+                assert gd <= group_delay_ms, \
+                    f'Given group delay {group_delay_ms*1000} ms is lower ' +\
+                    f'than minimal group delay {gd*1000} ms for channel {n}'
+                gd = group_delay_ms
+        else:
+            gd = group_delay_ms
+        lin_spectrum[:, n] = spectrum[:, n]*np.exp(
+            -1j * 2 * np.pi * f_vec * gd)
+    time_data = np.fft.irfft(lin_spectrum, axis=0)
+    sig_lin_phase = Signal(
+        None, time_data=time_data,
+        sampling_rate_hz=sampling_rate_hz, signal_type=signal_type)
+    return sig_lin_phase